Deduplicate SO3 rotation helpers
