+++ src/pyrecest/distributions/so3_bingham_distribution.py
@@ -15,13 +15,16 @@ from pyrecest.backend import (
     log,
     ndim,
     one_hot,
-    stack,
     sum,
     transpose,
     zeros,
 )
 
-from ._so3_helpers import geodesic_distance, normalize_quaternions
+from ._so3_helpers import (
+    geodesic_distance,
+    normalize_quaternions,
+    quaternions_to_rotation_matrices,
+)
 from .hypersphere_subset.bingham_distribution import BinghamDistribution
 from .hypersphere_subset.hyperhemispherical_bingham_distribution import (
     HyperhemisphericalBinghamDistribution,
@@ -78,7 +81,7 @@ class SO3BinghamDistribution(HyperhemisphericalBinghamDistribution):
     @staticmethod
     def _orthogonal_completion(mode):
         mode = SO3BinghamDistribution._canonicalize_quaternion(mode)
-        basis = []
+        basis: list = []
         for column in eye(4):
             vector = column - dot(column, mode) * mode
             for existing in basis:
@@ -225,26 +228,7 @@ class SO3BinghamDistribution(HyperhemisphericalBinghamDistribution):
     @staticmethod
     def as_rotation_matrices(quaternions):
         """Convert scalar-last quaternions to rotation matrices."""
-        quaternions = SO3BinghamDistribution._normalize_quaternions(quaternions)
-        x, y, z, w = (
-            quaternions[:, 0],
-            quaternions[:, 1],
-            quaternions[:, 2],
-            quaternions[:, 3],
-        )
-        row_0 = stack(
-            (1.0 - 2.0 * (y * y + z * z), 2.0 * (x * y - z * w), 2.0 * (x * z + y * w)),
-            axis=-1,
-        )
-        row_1 = stack(
-            (2.0 * (x * y + z * w), 1.0 - 2.0 * (x * x + z * z), 2.0 * (y * z - x * w)),
-            axis=-1,
-        )
-        row_2 = stack(
-            (2.0 * (x * z - y * w), 2.0 * (y * z + x * w), 1.0 - 2.0 * (x * x + y * y)),
-            axis=-1,
-        )
-        return stack((row_0, row_1, row_2), axis=-2)
+        return quaternions_to_rotation_matrices(quaternions)
 
     def mean_rotation_matrix(self):
         """Return the modal/principal rotation matrix."""
