+++ src/pyrecest/distributions/so3_tangent_gaussian_distribution.py
@@ -25,7 +25,12 @@ from pyrecest.backend import (
     zeros,
 )
 
-from ._so3_helpers import as_batch, geodesic_distance, normalize_quaternions
+from ._so3_helpers import (
+    as_batch,
+    geodesic_distance,
+    normalize_quaternions,
+    quaternions_to_rotation_matrices,
+)
 from .abstract_bounded_domain_distribution import AbstractBoundedDomainDistribution
 from .nonperiodic.gaussian_distribution import GaussianDistribution
 
@@ -140,26 +145,7 @@ class SO3TangentGaussianDistribution(AbstractBoundedDomainDistribution):
     @staticmethod
     def as_rotation_matrices(quaternions):
         """Convert scalar-last quaternions to rotation matrices."""
-        quaternions = SO3TangentGaussianDistribution._normalize_quaternions(quaternions)
-        x, y, z, w = (
-            quaternions[:, 0],
-            quaternions[:, 1],
-            quaternions[:, 2],
-            quaternions[:, 3],
-        )
-        row_0 = stack(
-            (1.0 - 2.0 * (y * y + z * z), 2.0 * (x * y - z * w), 2.0 * (x * z + y * w)),
-            axis=-1,
-        )
-        row_1 = stack(
-            (2.0 * (x * y + z * w), 1.0 - 2.0 * (x * x + z * z), 2.0 * (y * z - x * w)),
-            axis=-1,
-        )
-        row_2 = stack(
-            (2.0 * (x * z - y * w), 2.0 * (y * z + x * w), 1.0 - 2.0 * (x * x + y * y)),
-            axis=-1,
-        )
-        return stack((row_0, row_1, row_2), axis=-2)
+        return quaternions_to_rotation_matrices(quaternions)
 
     def pdf(self, xs):
         """Evaluate the tangent Gaussian density at SO(3) quaternions."""
